+++ setup/dashboard/inc/info.bw_page.php
@@ -1,8 +1,10 @@
 <?php
 
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 $bw_pages = [
-    ['key' => 's', 'title' => 'Top 10 days'],
-    ['key' => 'h', 'title' => 'Last 24 hours'],
+    ['key' => 't', 'title' => 'Top 10 days'],
+    ['key' => 'h', 'title' => 'Recent hours'],
     ['key' => 'd', 'title' => 'Last 30 days'],
     ['key' => 'm', 'title' => 'Last 12 months'],
 ];