+++ setup/dashboard/inc/info.package.php
@@ -1,5 +1,7 @@
 <?php
 
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 require_once($_SERVER['DOCUMENT_ROOT'].'/inc/util.php');
 
 $username = getMaster();
@@ -110,6 +112,20 @@ $packageList = [
             ],
         ],
     ], [
+        'package'     => 'emby',
+        'name'        => 'Emby',
+        'description' => 'EMBY',
+        'lockfile'    => '/install/.emby.lock',
+        'uninstall'   => 'UNINSTALL_EMBY_TXT',
+        'boxonly'     => false,
+        'services'    => [
+            'emby-server' => [
+                'process'  => 'emby-server',
+                'name'     => 'Emby',
+                'username' => 'emby',
+            ],
+        ],
+    ], [
         'package'     => 'fail2ban',
         'name'        => 'Fail2ban',
         'description' => 'FAIL2BAN',
@@ -146,7 +162,7 @@ $packageList = [
         'boxonly'     => false,
         'services'    => [
             'filebrowser-ee' => [
-                'process'  => 'filebrowser-ee',
+                'process'  => 'filebrowser',
                 'name'     => 'File Browser Enhanced',
                 'username' => $username,
             ],
@@ -177,6 +193,27 @@ $packageList = [
                 'process'  => 'flood',
                 'name'     => 'Flood',
                 'username' => $username,
+            ],
+        ],
+    ], [
+        'package'     => 'h5ai',
+        'name'        => 'h5ai',
+        'description' => 'H5AI',
+        'lockfile'    => '/install/.h5ai.lock',
+        'install'     => 'BOX_TOOLTIP_H5AI',
+        'boxonly'     => true,
+    ], [
+        'package'     => 'jellyfin',
+        'name'        => 'Jellyfin',
+        'description' => 'JELLYFIN',
+        'lockfile'    => '/install/.jellyfin.lock',
+        'uninstall'   => 'UNINSTALL_JELLYFIN_TXT',
+        'boxonly'     => false,
+        'services'    => [
+            'jellyfin' => [
+                'process'  => 'jellyfin',
+                'name'     => 'Jellyfin',
+                'username' => 'jellyfin',
             ],
         ],
     ], [
@@ -257,6 +294,20 @@ $packageList = [
             ],
         ],
     ], [
+        'package'     => 'rclone',
+        'name'        => 'Rclone',
+        'description' => 'RCLONE',
+        'lockfile'    => "/install/.{$username}.rclone.lock",
+        'uninstall'   => 'UNINSTALL_RCLONE_TXT',
+        'boxonly'     => false,
+        'services'    => [
+            'rclone-web' => [
+                'process'  => 'rcd',
+                'name'     => 'Rclone',
+                'username' => $username,
+            ],
+        ],
+    ], [
         'package'     => 'rtorrent',
         'name'        => 'rTorrent',
         'description' => 'RTORRENT',
@@ -279,6 +330,20 @@ $packageList = [
         'lockfile'    => '/install/.rutorrent.lock',
         'uninstall'   => 'UNINSTALL_RUTORRENT_TXT',
         'boxonly'     => false,
+    ], [
+        'package'     => 'sabnzbd',
+        'name'        => 'SABnzbd',
+        'description' => 'SABNZBD',
+        'lockfile'    => "/install/.{$username}.sabnzbd.lock",
+        'uninstall'   => 'UNINSTALL_SABNZBD_TXT',
+        'boxonly'     => false,
+        'services'    => [
+            'sabnzbd' => [
+                'process'  => 'sabnzbd',
+                'name'     => 'SABnzbd',
+                'username' => $username,
+            ],
+        ],
     ], [
         'package'     => 'speedtest',
         'name'        => 'SpeedTest',
@@ -329,6 +394,20 @@ $packageList = [
             ],
         ],
     ], [
+        'package'     => 'qbclientblocker',
+        'name'        => 'qBittorrent-ClientBlocker',
+        'description' => 'QBITTORRENTCLIENTBLOCKER',
+        'lockfile'    => "/install/.{$username}.qbittorrent-clientblocker.lock",
+        'uninstall'   => 'UNINSTALL_QBITTORRENTCLIENTBLOCKER_TXT',
+        'boxonly'     => false,
+        'services'    => [
+            'qbittorrent-clientblocker' => [
+                'process'  => 'qBittorrent-ClientBlocker',
+                'name'     => 'qBittorrent-ClientBlocker',
+                'username' => $username,
+            ],
+        ],
+    ], [
         'package'     => 'x2go',
         'name'        => 'x2go',
         'description' => 'X2GO',
@@ -350,6 +429,20 @@ $packageList = [
             ],
         ],
     ], [
+        'package'     => 'webdav',
+        'name'        => 'WebDAV',
+        'description' => 'WEBDAV',
+        'lockfile'    => "/install/.{$username}.webdav.lock",
+        'uninstall'   => 'UNINSTALL_WEBDAV_TXT',
+        'boxonly'     => false,
+        'services'    => [
+            'webdav' => [
+                'process'  => 'webdav',
+                'name'     => 'WebDAV',
+                'username' => $username,
+            ],
+        ],
+    ], [
         'skip'     => true,
         'package'  => 'vsftpd',
         'name'     => 'vsFTPD',
@@ -363,14 +456,14 @@ $packageList = [
         ],
     ], [
         'skip'     => true,
-        'package'  => 'shellinabox',
+        'package'  => 'ttyd',
         'name'     => 'Web Console',
-        'lockfile' => '/install/.shellinabox.lock',
+        'lockfile' => '/install/.ttyd.lock',
         'services' => [
-            'shellinabox' => [
-                'process'  => 'shellinabox',
+            'ttyd' => [
+                'process'  => 'ttyd',
                 'name'     => 'Web Console',
-                'username' => 'shellinabox',
+                'username' => $username,
             ],
         ],
     ],
@@ -381,98 +474,137 @@ foreach ($packageList as $package) {
     if (array_key_exists($package['package'], $packageMap)) {
         error_log("package '{$package['package']}' duplicated in package list!", 0);
     }
-    $packageMap[$package['package']] = $package;
+    $packageName              = (string) $package['package'];
+    $packageMap[$packageName] = $package;
+}
+
+/**
+ * @param string $packageName
+ *
+ * @return array<string,mixed>|null
+ */
+function getPackage($packageName) {
+    global $packageMap;
+    if (array_key_exists($packageName, $packageMap)) {
+        return $packageMap[$packageName];
+    }
+
+    return null;
 }
 
 $menuList = [
     [
         'name'    => 'ruTorrent',
         'service' => true,
-        'ref'     => $packageMap['rutorrent'],
+        'ref'     => getPackage('rutorrent'),
         'url'     => '/rutorrent/',
         'logo'    => 'img/brands/rtorrent.png',
     ], [
         'name'    => 'Flood',
         'service' => true,
-        'ref'     => $packageMap['flood'],
+        'ref'     => getPackage('flood'),
         'url'     => "/{$username}/flood/",
         'logo'    => 'img/brands/flood.png',
     ], [
         'name'    => 'Deluge Web',
         'service' => true,
-        'ref'     => $packageMap['deluge'],
+        'ref'     => getPackage('deluge'),
         'url'     => '/deluge/',
         'logo'    => 'img/brands/deluge.png',
     ], [
         'name'    => 'Transmission Web Control',
         'service' => true,
-        'ref'     => $packageMap['transmission'],
+        'ref'     => getPackage('transmission'),
         'url'     => '/transmission',
         'logo'    => 'img/brands/transmission.png',
     ], [
         'name'    => 'qBittorrent',
         'service' => true,
-        'ref'     => $packageMap['qbittorrent'],
+        'ref'     => getPackage('qbittorrent'),
         'url'     => '/qbittorrent/',
         'logo'    => 'img/brands/qbittorrent.png',
     ], [
         'name'    => 'BTSync',
         'service' => true,
-        'ref'     => $packageMap['btsync'],
+        'ref'     => getPackage('btsync'),
         'url'     => "/{$username}.btsync/",
         'logo'    => 'img/brands/btsync.png',
     ], [
+        'name'    => 'Emby',
+        'service' => true,
+        'ref'     => getPackage('emby'),
+        'url'     => '/emby/',
+        'logo'    => 'img/brands/emby.png',
+    ], [
         'name'    => 'File Browser',
         'service' => true,
-        'ref'     => $packageMap['filebrowser'],
+        'ref'     => getPackage('filebrowser'),
         'url'     => '/filebrowser/',
         'logo'    => 'img/brands/filebrowser.png',
     ], [
         'name'    => 'File Browser Enhanced',
         'service' => true,
-        'ref'     => $packageMap['filebrowser-ee'],
+        'ref'     => getPackage('filebrowser-ee'),
         'url'     => '/filebrowser-ee/',
         'logo'    => 'img/brands/filebrowser.png',
     ], [
         'name'    => 'FlexGet',
         'service' => true,
-        'ref'     => $packageMap['flexget'],
+        'ref'     => getPackage('flexget'),
         'url'     => '/flexget/',
         'logo'    => 'img/brands/flexget.png',
     ], [
+        'name'    => 'Jellyfin',
+        'service' => true,
+        'ref'     => getPackage('jellyfin'),
+        'url'     => '/jellyfin/',
+        'logo'    => 'img/brands/jellyfin.png',
+    ], [
         'name'    => 'NetData',
         'service' => true,
-        'ref'     => $packageMap['netdata'],
+        'ref'     => getPackage('netdata'),
         'url'     => '/netdata/',
         'logo'    => 'img/brands/netdata.png',
     ], [
         'name'    => 'noVNC',
         'service' => true,
-        'ref'     => $packageMap['novnc'],
+        'ref'     => getPackage('novnc'),
         'url'     => '/vnc/',
         'logo'    => 'img/brands/novnc.png',
     ], [
         'name'    => 'Plex',
         'service' => true,
-        'ref'     => $packageMap['plex'],
+        'ref'     => getPackage('plex'),
         'url'     => '/web/',
         'logo'    => 'img/brands/plex.png',
     ], [
+        'name'    => 'Rclone',
+        'service' => true,
+        'ref'     => getPackage('rclone'),
+        'url'     => '/rclone/',
+        'logo'    => 'img/brands/rclone.png',
+    ], [
+        'name'    => 'SABnzbd',
+        'service' => true,
+        'ref'     => getPackage('sabnzbd'),
+        'url'     => '/sabnzbd',
+        'logo'    => 'img/brands/sabnzbd.png',
+    ], [
         'name'    => 'SpeedTest',
         'service' => false,
-        'ref'     => $packageMap['speedtest'],
+        'ref'     => getPackage('speedtest'),
         'url'     => '/speedtest/',
         'logo'    => 'img/brands/speedtest.png',
     ], [
         'name'    => 'Syncthing',
         'service' => true,
-        'ref'     => $packageMap['syncthing'],
+        'ref'     => getPackage('syncthing'),
         'url'     => "/{$username}.syncthing/",
         'logo'    => 'img/brands/syncthing.png',
     ], [
         'name'    => 'ZNC',
         'service' => true,
-        'ref'     => $packageMap['znc'],
+        'ref'     => getPackage('znc'),
         'url'     => '/znc/',
         'logo'    => 'img/brands/znc.png',
     ],
@@ -481,19 +613,19 @@ $menuList = [
 $downloadList = [
     [
         'name' => 'rTorrent',
-        'ref'  => $packageMap['rtorrent'],
+        'ref'  => getPackage('rtorrent'),
         'url'  => "/{$username}.rtorrent.downloads",
     ], [
         'name' => 'Deluge',
-        'ref'  => $packageMap['deluge'],
+        'ref'  => getPackage('deluge'),
         'url'  => "/{$username}.deluge.downloads",
     ], [
         'name' => 'Transmission',
-        'ref'  => $packageMap['transmission'],
+        'ref'  => getPackage('transmission'),
         'url'  => "/{$username}.transmission.downloads",
     ], [
         'name' => 'qBittorrent',
-        'ref'  => $packageMap['qbittorrent'],
+        'ref'  => getPackage('qbittorrent'),
         'url'  => "/{$username}.qbittorrent.downloads",
     ], [
         'name'     => 'OpenVPN Config',
@@ -556,12 +688,12 @@ function __check_package_config($package) {
     assert(array_key_exists('package', $package));
     assert(array_key_exists('name', $package));
     $skip = array_key_exists('skip', $package) ? $package['skip'] : false;
-    if (!$skip) {
+    if ($skip === false) {
         assert(array_key_exists('description', $package));
         assert(array_key_exists('lockfile', $package));
         assert(array_key_exists('boxonly', $package));
         $boxonly = $package['boxonly'];
-        if ($boxonly) {
+        if ($boxonly === true) {
             assert(array_key_exists('install', $package));
         } else {
             assert(array_key_exists('uninstall', $package));