+++ setup/dashboard/widgets/load.php
@@ -1,5 +1,0 @@
-<?php
-
-require_once($_SERVER['DOCUMENT_ROOT'].'/inc/info.system.php');
-
-echo SystemInfo::loadavg();