+++ setup/dashboard/ws/src/handler/log.ts
@@ -1,9 +1,9 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 import Constant from "../constant";
 import { Socket } from "socket.io";
 
 
-// eslint-disable-next-line no-unused-vars
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export default (client: Socket, next?: (err?: Error) => void) => {
     const clientIp = client.handshake.headers["x-forwarded-for"] || client.handshake.address;
     console.log(`${client.id} connect from ${clientIp}`);